+++ wireform-websocket/test/Spec.hs
@@ -4,11 +4,13 @@ import Test.Tasty (defaultMain, testGroup)
 
 import qualified Test.Frame
 import qualified Test.Handshake
+import qualified Test.URI
 import qualified Test.Echo
 
 main :: IO ()
 main = defaultMain $ testGroup "wireform-websocket"
   [ Test.Frame.tests
   , Test.Handshake.tests
+  , Test.URI.tests
   , Test.Echo.tests
   ]